api controller refactor to subdir
