+++ lib/oli_web/controllers/api/attempt_controller.ex
@@ -1,4 +1,4 @@
-defmodule OliWeb.AttemptController do
+defmodule OliWeb.Api.AttemptController do
   use OliWeb, :controller
   use OpenApiSpex.Controller
