+++ lib/oli_web/controllers/api/media_controller.ex
@@ -1,4 +1,4 @@
-defmodule OliWeb.MediaController do
+defmodule OliWeb.Api.MediaController do
   @moduledoc tags: ["Media Library Service"]
 
   @moduledoc """
